+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Security - in case of vulnerabilities.
 - [x] Added `dueTime` to `InvoiceService::issue()` 
 - [x] Added new properties to `Transaction`: `requestAmount`, `requestCurrency`, `purchaseAmount`, `purchaseCurrency`, `isProcessedOutside`
 - [x] Added field `additionalCriteria` into `GatewayAccount` object
+- [x] Added field `dueTimeShift` into `BillingAnchor` object
 
 ### Deprecated
 - [x] Deprecated `Transaction` entity method: `getPaymentCardId`